Work towards getting issue access.
Note:
Please god can I just get a easy way to tell the json serializer what to convert specific interfaces to?????

Like please, I want you to convert the json to a IUserReference which only contains one property, but since it is a interface, you have no idea how to do it. So WHY don't you let me easily tell you that if you find a IUserReference then you create a object of type UserReference????

Is that rocketscience?